사업장 _ 문제해결 (김현송 팀장님 의뢰건)

+++ FamTec/FamTec/FamTec/Shared/Server/DTO/Store/PeriodicInventoryRecordDTO.cs
@@ -46,7 +46,7 @@
         /// <summary>
         /// 총 재고 수량
         /// </summary>
-        public long TotalStockNum { get; set; }
+        public int TotalStockNum { get; set; }
 
         /// <summary>
         /// 이월재고 수량
